fix: moved cli modules to separate directory
