+++ src/cli/preprocessors/resize/fit.rs
